% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta1_sim_JM}
\alias{beta1_sim_JM}
\title{function to simulate beta1 for joint model}
\usage{
beta1_sim_JM(eta, intercept1, W1, kappa_eta, Z1, hyper_fixed)
}
\arguments{
\item{eta}{}

\item{intercept1}{}

\item{W1}{}

\item{kappa_eta}{}

\item{Z1}{}
}
\description{
function to simulate beta1 for joint model
}
