% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta2_sim_JM}
\alias{beta2_sim_JM}
\title{Simulate Coefficients \code{beta2} of \code{mu} for Joint Model}
\usage{
beta2_sim_JM(mu, intercept2, beta, kappa_mu, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the response variable for the second part of the model.}

\item{intercept2}{A numeric value representing the intercept for the second part of the model.}

\item{beta}{A numeric value representing the coefficient for the first part of the model.}

\item{kappa_mu}{A numeric value representing the precision parameter associated with \code{mu}.}

\item{W1}{A numeric vector representing the latent random effects for the first part of the model.}

\item{W2}{A numeric vector representing the latent random effects for the second part of the model.}

\item{Z2}{A numeric matrix of covariates corresponding to the response variable \code{mu}.}

\item{hyper_fixed}{A list containing hyperparameters, specifically the prior precision for \code{beta2}.}
}
\value{
A numeric vector of simulated values for \code{beta2} from its full conditional distribution.
}
\description{
This function simulates the coefficients \code{beta2} for a joint model based on its full conditional distribution, which incorporates the effects of covariates and random effects.
}
\examples{
# Example data for the joint model
mu <- rnorm(100)
intercept2 <- 0.5
beta <- 1
kappa_mu <- 2
W1 <- rnorm(100)
W2 <- rnorm(100)
Z2 <- matrix(rnorm(500), 100, 5)
hyper_fixed <- list(beta2 = 1)

# Simulate beta2 for joint model
beta2_sim <- beta2_sim_JM(mu, intercept2, beta, kappa_mu, W1, W2, Z2, hyper_fixed)
}
