% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta2_sim_thr}
\alias{beta2_sim_thr}
\title{Simulate the Intercept Parameter for Threshold Model}
\usage{
beta2_sim_thr(mu, intercept2, beta, kappa_mu, W2, Z2, Z2.crossprd, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the current state of the latent mean parameters.}

\item{kappa_mu}{A numeric value representing the precision of the mean effects in the threshold model.}

\item{W2}{A numeric vector representing the random effects for the model.}

\item{Z2}{A matrix of covariates, where each row corresponds to an observation.}

\item{hyper_fixed}{A list containing the hyperparameter for \code{intercept2}, which includes its prior precision (\code{hyper_fixed$intercept2}).}

\item{beta2}{A numeric vector of coefficients for the covariate matrix \code{Z2}.}
}
\value{
A simulated value for \code{intercept2} from its full conditional normal distribution.
}
\description{
This function simulates the parameter \code{intercept2}, which is the intercept term in a threshold model, based on the full conditional distribution derived from the data and model parameters.
}
\examples{
# Sample parameters
mu <- rnorm(100)
beta2 <- rnorm(5)
W2 <- rnorm(100)
kappa_mu <- 2
Z2 <- matrix(rnorm(500), ncol=5)
hyper_fixed <- list(intercept2 = 0.1)

# Simulate intercept2 for threshold model
intercept2_sim <- intercept2_sim_thr(mu, beta2, W2, kappa_mu, Z2, hyper_fixed)

}
