% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta2_sim_thr_ind}
\alias{beta2_sim_thr_ind}
\title{Function to simulate beta2  for threshold indicator models}
\usage{
beta2_sim_thr_ind(
  mu,
  intercept2,
  beta,
  kappa_mu,
  W2,
  Z2,
  Z2.crossprd,
  hyper_fixed
)
}
\arguments{
\item{mu}{}

\item{intercept2}{}

\item{beta}{}

\item{kappa_mu}{}

\item{W2}{}

\item{Z2}{}

\item{Z2.crossprd}{}
}
\description{
Function to simulate beta2  for threshold indicator models
}
