% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta_sim_JM}
\alias{beta_sim_JM}
\title{Function to simulate the beta for joint model}
\usage{
beta_sim_JM(mu, intercept2, kappa_mu, beta2, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{intercept2}{}

\item{kappa_mu}{}

\item{beta2}{}

\item{W1}{}

\item{W2}{}

\item{Z2}{}
}
\description{
Function to simulate the beta for joint model
}
