% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_updates.R
\name{block_rw_update}
\alias{block_rw_update}
\title{Univariate (block) updates based on random walk proposals}
\usage{
block_rw_update(
  par_curr,
  par_name,
  loglik,
  logprior,
  var_markdist,
  lb = -Inf,
  ub = Inf,
  transform = FALSE,
  ...
)
}
\arguments{
\item{par_curr}{current parameter values}

\item{par_name}{name of the parameter of interest. While writing the log-likelihood and log-prior the parameter of interest should have same parameter values}

\item{loglik}{log-likelihood function}

\item{logprior}{log-prior function}

\item{var_markdist}{variance of the random walk proposals}

\item{lb}{lower support of the parameter of interest}

\item{ub}{upper support of the parameter of interest}

\item{transform}{logical, if TRUE transformation are applied are to convert the current parameter space to whole real line and then performed Gaussian proposals. If FALSE then a truncated normal proposals are used in the (lb, ub)}

\item{...}{additional parameter}
}
\value{
return an updated values of parameter of interest
}
\description{
Univariate (block) updates based on random walk proposals
}
