% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{cond.pbGPD}
\alias{cond.pbGPD}
\title{Conditional Cumulative Distribution Function of Mixture of Truncated Beta and GPD}
\usage{
cond.pbGPD(x, mu, u.thr, k, xi, sigma, sigma.GP, thr.prob)
}
\arguments{
\item{x}{A numeric value representing the high quantile at which the CDF is evaluated. It should be greater than \code{u.thr}.}

\item{mu}{A numeric vector representing the location parameter of the mixture distribution.}

\item{u.thr}{A positive numeric value specifying the threshold separating the truncated beta and GPD components.}

\item{k}{A positive numeric value representing the shape1 parameter of the truncated beta distribution.}

\item{xi}{A numeric value representing the shape parameter of the GPD.}

\item{sigma}{A numeric vector specifying the scale parameter for the truncated beta distribution; must be of the same length as \code{mu}.}

\item{sigma.GP}{A positive numeric value representing the scale parameter of the GPD.}

\item{thr.prob}{A numeric value between 0 and 1 indicating the mixture probability for the GPD component (i.e., the probability that a sample exceeds the threshold \code{u.thr}).}

\item{log.p}{A logical value; if \code{TRUE}, the log of the CDF is returned. If \code{FALSE}, the CDF is returned.}
}
\value{
A numeric vector of the same length as \code{mu}, containing the conditional CDF values (or log CDF values) evaluated at \code{x}.
}
\description{
This function computes the conditional cumulative distribution function (CDF) for a mixture distribution, which includes a truncated beta distribution for values below the threshold \code{u.thr} and a Generalized Pareto Distribution (GPD) for values above \code{u.thr}.
}
\details{
This function provides a conditional CDF for a mixture distribution:
\itemize{
\item For \code{x} greater than \code{u.thr}, the CDF is derived from the GPD component.
\item For \code{x} less than or equal to \code{u.thr}, the CDF is derived from the truncated beta distribution.
The value of \code{mu} is used to conditionally determine which distribution is applicable, and \code{thr.prob} specifies the probability of being in the GPD component.
}
}
\examples{
# Example usage of cond.pbGPD function
x <- 8
mu <- seq(log(1), log(20), length.out = 100)
u.thr <- 5
k <- 2
xi <- 0.5
sigma <- rep(1, length(mu))
sigma.GP <- 1.5
thr.prob <- 0.4
cdf <- cond.pbGPD(x, mu, u.thr, k, xi, sigma, sigma.GP, thr.prob)
plot(mu, cdf, type = "l", main = "Conditional CDF of Mixture Distribution")
}
\seealso{
\code{\link{pbetat}} for the truncated beta distribution CDF function,
\code{\link[evd]{pgpd}} for the Generalized Pareto Distribution CDF function.
}
