% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{cond.ptgGPD}
\alias{cond.ptgGPD}
\title{Title}
\usage{
cond.ptgGPD(x, mu, u.thr, k, xi, sigma, sigma.GP, thr.prob)
}
\arguments{
\item{x}{}

\item{mu}{}

\item{u.thr}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{thr.prob}{}
}
\description{
Title
}
