% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dbetat}
\alias{dbetat}
\title{density function of beta distribution with support in (0, c)}
\usage{
dbetat(x, u.thr, shape1, shape2, log = TRUE)
}
\arguments{
\item{x}{}

\item{shape1}{}

\item{shape2}{}

\item{log}{}
}
\description{
density function of beta distribution with support in (0, c)
}
