% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dbGPD}
\alias{dbGPD}
\title{Density function of mixture of beta in (0, u.thr)  and GPD}
\usage{
dbGPD(x, u.thr, k, xi, sigma, sigma.GP, thr.prob, log = TRUE)
}
\arguments{
\item{x}{a vector of observations}

\item{u.thr}{}

\item{k}{shape1 parameter for the  beta distribution}

\item{xi}{shape parameter of GP}

\item{sigma}{a vectors of scale parameters with same length as the length of x}

\item{thr.prob}{}

\item{log}{}
}
\description{
Density function of mixture of beta in (0, u.thr)  and GPD
}
