% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dgammat}
\alias{dgammat}
\title{Density of the Truncated Gamma Distribution}
\usage{
dgammat(x, upper.bound, shape, scale, log = TRUE)
}
\arguments{
\item{x}{A numeric vector of quantiles at which the density is evaluated.}

\item{upper.bound}{A numeric value indicating the upper truncation bound for the Gamma distribution.}

\item{shape}{A positive numeric value representing the shape parameter of the Gamma distribution.}

\item{scale}{A positive numeric value representing the scale parameter of the Gamma distribution.}

\item{log}{Logical; if \code{TRUE}, the log-density is returned. If \code{FALSE}, the density is returned. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of density values for the truncated Gamma distribution at the specified quantiles \code{x}.
}
\description{
This function calculates the density of a Gamma distribution truncated to an upper bound.
It provides the log-density or the density itself based on the user's input.
}
\details{
The Gamma distribution is truncated at an upper bound specified by \code{upper.bound}.
The function normalizes the density by the probability that the Gamma variable lies within the range from zero to \code{upper.bound}.
If \code{log = TRUE}, the function returns the log-density; otherwise, it returns the density.
}
\examples{
# Calculate the density for a sequence of quantiles
x <- seq(0, 5, by = 0.1)
shape <- 2
scale <- 1
upper.bound <- 5
density_values <- dgammat(x, upper.bound, shape, scale, log = FALSE)
log_density_values <- dgammat(x, upper.bound, shape, scale, log = TRUE)

}
\seealso{
\code{\link{dgamma}} for the standard Gamma density function.
}
