% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dtnorm}
\alias{dtnorm}
\title{Density of a univariate truncated Normal}
\usage{
dtnorm(x, mean = 0, sd = 1, a = -Inf, b = Inf, log = FALSE)
}
\arguments{
\item{x}{\link{numeric} observation vector}

\item{mean}{\link{numeric} scalar location parameter}

\item{sd}{\link{numeric} scalar scale parameter}

\item{a}{\link{numeric} lower bound}

\item{b}{\link{numeric} upper bound}

\item{log}{\link{logical} if \code{TRUE}, return the log density}
}
\description{
Density of a univariate truncated Normal
}
