% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_GPD_param}
\alias{gamma_prior_GPD_param}
\title{Title}
\usage{
gamma_prior_GPD_param(cur_par, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{}

\item{hyper.mu_fixed}{}

\item{mark_dist}{}
}
\description{
Title
}
