% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_markdist_JM}
\alias{gamma_prior_markdist_JM}
\title{prior distributions for the mark distributions}
\usage{
gamma_prior_markdist_JM(cur_par, mark_dist, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{}

\item{mark_dist}{}

\item{hyper.mu_fixed}{}
}
\description{
prior distributions for the mark distributions
}
