% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_markdist_thr}
\alias{gamma_prior_markdist_thr}
\title{Prior Distributions for Threshold Model's Distributions}
\usage{
gamma_prior_markdist_thr(cur_par, thr.family, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{A numeric vector representing the current parameter values of the mark distribution. For both gamma and log-normal families, it typically includes the shape parameter \code{k}.}

\item{thr.family}{A character string indicating the family of the threshold model. Can be either \code{"gamma"} or \code{"logNormal"}.}

\item{hyper.mu_fixed}{A numeric vector of length 2, providing the shape and rate hyperparameters of the gamma prior distribution for the mark distribution parameter.}
}
\value{
A numeric value representing the log-prior density of the parameter under the specified threshold family.
}
\description{
Computes the log prior density for the parameter of the distribution in a threshold model, supporting both gamma and log-normal families.
}
\details{
This function calculates the log-prior density for the shape parameter \code{k} of the mark distribution based on the specified threshold family:
\itemize{
\item If \code{thr.family = "gamma"}, the prior for \code{k} is a gamma distribution with the specified hyperparameters.
\item If \code{thr.family = "logNormal"}, the same gamma prior is applied to \code{k}.
}
}
\examples{
# Gamma threshold model
cur_par <- c(2) # Current parameter value (shape k)
thr.family <- "gamma"
hyper.mu_fixed <- c(1, 0.5) # Hyperparameters for the gamma prior
gamma_prior_markdist_thr(cur_par, thr.family, hyper.mu_fixed)

# Log-normal threshold model
cur_par <- c(3) # Current parameter value (shape k)
thr.family <- "logNormal"
hyper.mu_fixed <- c(2, 1) # Hyperparameters for the gamma prior
gamma_prior_markdist_thr(cur_par, thr.family, hyper.mu_fixed)
}
