% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-missing locations.R
\name{impute.NA.A}
\alias{impute.NA.A}
\title{Impute NA Values for Landslide sizes}
\usage{
impute.NA.A(
  CV,
  ind_NA_A,
  ind_zeros_counts,
  mu,
  thr.prob,
  cur_par,
  mark_dist,
  threshold,
  thr.exceed.ind = NULL
)
}
\arguments{
\item{CV}{A character string specifying the cross-validation type. Can be \code{"WS"} for within-sample or \code{"OOS"} for
out-of-sample.}

\item{ind_NA_A}{A logical vector indicating which elements in the accumulation data are missing (NA).}

\item{ind_zeros_counts}{A logical vector indicating which elements in the count data are zeros.}

\item{mu}{A numeric vector of linear predictors, typically on the log scale.}

\item{thr.prob}{A numeric value indicating the threshold exceedance probability.}

\item{cur_par}{A numeric vector of hyper parameters for the respective distribution (e.g., \code{k}, \code{xi}).}

\item{mark_dist}{A character string indicating the marginal distribution family. Accepts \code{"eGPD"}, \code{"bGPD"}, or \code{"tgGPD"}.}

\item{threshold}{A numeric vector representing the threshold for the generalized Pareto distribution.}

\item{thr.exceed.ind}{(Optional) A logical vector indicating if the threshold is exceeded. Default is \code{NULL}.}
}
\value{
A list containing:
\item{thr.exceed.ind}{A logical vector indicating whether the threshold is exceeded.}
\item{imputed_NA_A}{A numeric vector of imputed values for the missing accumulations.}
}
\description{
This function imputes missing (NA) values in landslide sizes data based on the specified marginal distribution.
It handles different distribution families, including eGPD (extended Generalized Pareto Distribution), bGPD
(mixture of beta-GPD), and tgGPD (mixture of truncated gamma-GPD).
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The following is an example of the function usage and is not meant to be run directly.
CV <- "WS"
ind_NA_A <- c(TRUE, FALSE, TRUE)
ind_zeros_counts <- c(FALSE, TRUE, FALSE)
mu <- c(1.5, 0.8, 1.1)
thr.prob <- 0.9
cur_par <- c(0.5, 1.2, 0.7)
mark_dist <- "eGPD"
threshold <- c(0.6, 1.0, 0.8)
impute.NA.A(CV, ind_NA_A, ind_zeros_counts, mu, thr.prob, cur_par, mark_dist, threshold)
\dontshow{\}) # examplesIf}
}
