% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-misisng locations.R
\name{impute.NA.Y}
\alias{impute.NA.Y}
\title{Imputing NA for the landslides counts}
\usage{
impute.NA.Y(ind_NA_Y, eta, CV)
}
\arguments{
\item{ind_NA_Y}{}

\item{eta}{}

\item{CV}{}
}
\description{
Imputing NA for the landslides counts
}
