% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_all_other_param_indicator_model_probit_link}
\alias{init_fun_all_other_param_indicator_model_probit_link}
\title{Initialize Parameters for Indicator Model with Probit Link}
\usage{
init_fun_all_other_param_indicator_model_probit_link(Z2, A, seed)
}
\arguments{
\item{Z2}{A matrix of covariates for the model.}

\item{A}{A matrix (or data frame) representing additional covariates or transformations used in the model.}

\item{seed}{A numeric value used to set the seed for random number generation, ensuring reproducibility.}
}
\value{
A list containing:
\item{init.all.other.param}{A numeric vector of initialized values for model parameters, including kappas, intercepts, and covariate coefficients.}
\item{model.param.name.all.other.param}{A character vector or expression list representing the names of the parameters (e.g., \code{expression(kappa[w])}).}
}
\description{
This function generates initial values for the parameters of an indicator model, including coefficients for covariates and other model-specific parameters.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The following is an example usage of the function and is not meant to be run directly.
Z2 <- matrix(runif(10), nrow = 5, ncol = 2)
A <- matrix(runif(10), nrow = 5, ncol = 2)
seed <- 123
init_values <- init_fun_all_other_param_indicator_model(Z2, A, seed)
\dontshow{\}) # examplesIf}
}
