% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_all_other_param_JM}
\alias{init_fun_all_other_param_JM}
\title{Initialize Values for Model Parameters in Joint Model}
\usage{
init_fun_all_other_param_JM(
  Z1,
  Z2,
  A,
  Y,
  seed,
  simulation,
  threshold,
  mark_dist,
  model_type,
  thr.acces.ind
)
}
\arguments{
\item{Z1}{A design matrix for covariates associated with the first set of parameters.}

\item{Z2}{A design matrix for covariates associated with the second set of parameters.}

\item{A}{A matrix representing spatial relationships or adjacency information.}

\item{Y}{A vector or matrix of response variables (e.g., landslide counts or sizes).}

\item{seed}{An integer to set the random seed for reproducibility.}

\item{simulation}{A logical indicating if the function is being used in a simulation setting.}

\item{threshold}{A numeric value representing the threshold for the model.}

\item{thr.acces.ind}{A vector of indices indicating which data points exceed the threshold.}
}
\value{
A list containing initial values of parameters required for the model.
}
\description{
Generates initial values for model parameters that are independent of the choice of mark distributions.
}
\examples{
# Example usage:
Z1 <- matrix(rnorm(100), 20, 5)
Z2 <- matrix(rnorm(200), 20, 10)
A <- matrix(1, nrow = 20, ncol = 20)
Y <- rpois(20, lambda = 5)
seed <- 123
simulation <- TRUE
threshold <- 3
thr.acces.ind <- which(Y > threshold)
initialize_parameters(Z1, Z2, A, Y, seed, simulation, threshold, thr.acces.ind)
}
