% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_all_other_param_thr_model}
\alias{init_fun_all_other_param_thr_model}
\title{Initialize All Other Parameters for Threshold Model}
\usage{
init_fun_all_other_param_thr_model(Z2, A, thr.family, seed, simulation)
}
\arguments{
\item{Z2}{A matrix of covariates for the model.}

\item{A}{A matrix (or data frame) representing additional covariates or transformations used in the model.}

\item{thr.family}{A character string specifying the family of the threshold model. Can be \code{"gamma"} or \code{"logNormal"}.}

\item{seed}{A numeric value used to set the seed for random number generation, ensuring reproducibility.}

\item{simulation}{A logical value indicating whether the function is being run in simulation mode.}
}
\value{
A list containing:
\item{init.all.other.param}{A numeric vector of initialized values for the model parameters, including kappas, intercepts, and covariate coefficients.}
\item{model.param.name.all.other.param}{A character vector or expression list representing the names of the parameters (e.g., \code{expression(kappa[w])}).}
}
\description{
This function generates initial values for the parameters of a threshold model based on the specified family and additional covariates. It handles both \code{"gamma"} and \code{"logNormal"} families.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The following is an example usage of the function and is not meant to be run directly.
Z2 <- matrix(runif(10), nrow = 5, ncol = 2)
A <- matrix(runif(10), nrow = 5, ncol = 2)
thr.family <- "gamma"
seed <- 123
simulation <- TRUE
init_values <- init_fun_all_other_param_thr_model(Z2, A, thr.family, seed, simulation)
\dontshow{\}) # examplesIf}
}
