% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_hyper.GP_JM}
\alias{init_fun_hyper.GP_JM}
\title{Initialize All Other Parameters for Joint Model}
\usage{
init_fun_hyper.GP_JM(mark_dist, seed)
}
\arguments{
\item{mark_dist}{A character string indicating the type of marginal distribution. Accepts \code{"eGPD"} for extended Generalized Pareto Distribution, \code{"bGPD"} for a mixture of beta-GPD, or \code{"tgGPD"} for a mixture of truncated gamma-GPD.}

\item{seed}{A numeric value to set the seed for random number generation to ensure reproducibility.}

\item{Z1}{A matrix of covariates for the first part of the model.}

\item{Z2}{A matrix of covariates for the second part of the model.}

\item{A}{A matrix (or data frame) representing additional covariates or transformations used in the model.}

\item{Y}{A vector of response variables for the model.}

\item{simulation}{A logical value indicating whether the function is being run in simulation mode.}

\item{threshold}{A numeric value or vector specifying the threshold for handling certain parameters in the model.}

\item{model_type}{A character string indicating the type of model. Can be \code{"FE"} for fixed effects or other specified model types.}

\item{thr.acces.ind}{A logical vector indicating whether each observation exceeds the threshold.}
}
\value{
A list containing:
\item{init.all.other.param}{A named list of initialized values for the model parameters, including kappas, intercepts, coefficients for covariates, and other model-specific parameters.}
\item{model.param.name.all.other.param}{An expression vector or character vector representing the names of the parameters for better display in plots or tables.}
}
\description{
This function generates initial values for the parameters of a joint model that are independent of the choice of marginal distributions. It handles initialization for fixed effects (FE) and other model types.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
Z1 <- matrix(runif(10), nrow = 5, ncol = 2)
Z2 <- matrix(runif(10), nrow = 5, ncol = 2)
A <- matrix(runif(10), nrow = 5, ncol = 2)
Y <- rnorm(5)
seed <- 123
simulation <- TRUE
threshold <- c(0.5, 0.8)
mark_dist <- "eGPD"
model_type <- "FE"
thr.acces.ind <- c(TRUE, FALSE, TRUE, FALSE, TRUE)
init_values <- init_fun_all_other_param_JM(Z1, Z2, A, Y, seed, simulation, threshold, mark_dist, model_type, thr.acces.ind)
\dontshow{\}) # examplesIf}
}
