% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_log.hyper.mu_thr_model}
\alias{init_fun_log.hyper.mu_thr_model}
\title{Initialize Parameters for Threshold Model}
\usage{
init_fun_log.hyper.mu_thr_model(thr.family, seed)
}
\arguments{
\item{thr.family}{A character string specifying the family of the threshold model. Can be \code{"gamma"} or \code{"logNormal"}.}

\item{seed}{A numeric value used to set the seed for random number generation, ensuring reproducibility.}
}
\value{
A list containing:
\item{log.hyper.mu.init}{A numeric vector representing the initialized value of the model parameter.}
\item{log.hyper.mu.name}{An expression object with the name of the parameter (e.g., \code{expression(k)}).}
}
\description{
This function generates initial values for the parameters of a threshold model based on the specified family.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The following is an example usage of the function and is not meant to be run directly.
thr.family <- "gamma"
seed <- 123
init_values <- init_fun_log.hyper.mu_thr_model(thr.family, seed)
\dontshow{\}) # examplesIf}
}
