% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept1_sim_JM}
\alias{intercept1_sim_JM}
\title{Simulate Intercept \code{intercept1} of \code{eta} for Joint Model}
\usage{
intercept1_sim_JM(eta, beta1, W1, kappa_eta, Z1, hyper_fixed)
}
\arguments{
\item{eta}{A numeric vector representing the response variable for the first part of the model.}

\item{beta1}{A numeric vector of coefficients for the covariates in \code{Z1}.}

\item{W1}{A numeric vector representing the latent random effects for the first part of the model.}

\item{kappa_eta}{A numeric value representing the precision parameter associated with \code{eta}.}

\item{Z1}{A numeric matrix of covariates corresponding to the response variable \code{eta}.}

\item{hyper_fixed}{A list containing hyperparameters, specifically the prior precision for \code{intercept1}.}
}
\value{
A simulated value for \code{intercept1} from its full conditional normal distribution.
}
\description{
This function simulates the intercept parameter \code{intercept1} for a joint model based on its full conditional distribution, which is normally distributed with precision derived from the latent variables and covariates.
}
\examples{
# Example data for the joint model
eta <- rnorm(100)
beta1 <- rnorm(5)
W1 <- rnorm(100)
kappa_eta <- 2
Z1 <- matrix(rnorm(500), 100, 5)
hyper_fixed <- list(intercept1 = 1.5)

# Simulate intercept1 for joint model
intercept1_sim <- intercept1_sim_JM(eta, beta1, W1, kappa_eta, Z1, hyper_fixed)
}
