% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept2_sim_JM}
\alias{intercept2_sim_JM}
\title{function to simulate the intercept2 for joint model}
\usage{
intercept2_sim_JM(mu, beta2, beta, W1, W2, kappa_mu, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{beta2}{}

\item{beta}{}

\item{W1}{}

\item{W2}{}

\item{kappa_mu}{}

\item{Z2}{}
}
\description{
function to simulate the intercept2 for joint model
}
