% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept2_sim_thr_ind}
\alias{intercept2_sim_thr_ind}
\title{Function to simulate the intercept2  for threshold indicator models}
\usage{
intercept2_sim_thr_ind(mu, beta2, W2, kappa_mu, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{beta2}{}

\item{W2}{}

\item{kappa_mu}{}

\item{Z2}{}

\item{beta}{}
}
\description{
Function to simulate the intercept2  for threshold indicator models
}
