% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{jac_inv_transfo}
\alias{jac_inv_transfo}
\title{Jacobean of the transformations made in \transfo}
\usage{
jac_inv_transfo(tpar, lb, ub, log = FALSE)
}
\arguments{
\item{tpar}{the transformed parameter}

\item{lb}{lower bound of parameter}

\item{ub}{upper bound of the parameter}

\item{log}{log=TRUE return log-Jacobean}
}
\description{
Jacobean of the transformations made in \transfo
}
