% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_eta_sim_JM}
\alias{kappa_eta_sim_JM}
\title{Function to simulate the kappa_eta for joint model}
\usage{
kappa_eta_sim_JM(eta, intercept1, beta1, W1, Z1, hyper_fixed)
}
\arguments{
\item{eta}{}

\item{intercept1}{}

\item{beta1}{}

\item{W1}{}

\item{Z1}{}
}
\description{
Function to simulate the kappa_eta for joint model
}
