% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_mu_sim_JM}
\alias{kappa_mu_sim_JM}
\title{Simulate Precision Parameter \code{kappa_mu} of \code{mu} for Joint Model}
\usage{
kappa_mu_sim_JM(mu, intercept2, beta2, beta, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the response variable.}

\item{intercept2}{A numeric value representing the intercept term in the model.}

\item{beta2}{A numeric vector of coefficients for covariates in \code{Z2}.}

\item{beta}{A numeric value representing the coefficient associated with random effect \code{W1}.}

\item{W1}{A numeric vector representing the first set of latent random effects in the model.}

\item{W2}{A numeric vector representing the second set of latent random effects in the model.}

\item{Z2}{A numeric matrix of covariates corresponding to the response variable.}

\item{hyper_fixed}{A list containing hyperparameters for the Gamma prior distribution, specifically the shape and rate parameters for \code{kappa_mu}.}
}
\value{
A simulated value for \code{kappa_mu} from the Gamma distribution.
}
\description{
This function simulates the precision parameter \code{kappa_mu} for a joint model from a full conditional Gamma distribution.
}
\examples{
# Example data for the joint model
mu <- rnorm(100)
intercept2 <- 0.5
beta2 <- rnorm(5)
beta <- 1.2
W1 <- rnorm(100)
W2 <- rnorm(100)
Z2 <- matrix(rnorm(500), 100, 5)
hyper_fixed <- list(kappa_mu = c(shape = 2, rate = 1))

# Simulate kappa_mu for joint model
kappa_sim <- kappa_mu_sim_JM(mu, intercept2, beta2, beta, W1, W2, Z2, hyper_fixed)
}
