% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_mu_sim_thr_ind_logit}
\alias{kappa_mu_sim_thr_ind_logit}
\title{Simulate the Precision Parameter \code{kappa_mu} for Threshold Indicator Models with logit link}
\usage{
kappa_mu_sim_thr_ind_logit(mu, intercept2, beta2, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the latent mean parameter of the model.}

\item{intercept2}{A numeric value indicating the intercept term in the model.}

\item{beta2}{A numeric vector representing the coefficients of covariates in the model.}

\item{W2}{A numeric vector indicating the latent (iCAR) random effects.}

\item{Z2}{A numeric matrix of covariates corresponding to the latent process \code{mu}.}

\item{hyper_fixed}{A list containing hyperparameters for the Gamma prior distribution, specifically the shape and rate parameters for \code{kappa_mu}.}
}
\value{
A simulated value for \code{kappa_mu} from the Gamma distribution.
}
\description{
This function simulates the precision parameter \code{kappa_mu} for threshold indicator models from a full conditional Gamma distribution.
}
\examples{
# Sample latent variables and covariates
mu <- rnorm(100)
intercept2 <- 0.5
beta2 <- rnorm(10)
W2 <- rnorm(100)
Z2 <- matrix(rnorm(1000), ncol = 10)
hyper_fixed <- list(kappa_mu = c(shape = 2, rate = 1))

# Simulate kappa_mu for threshold indicator model
kappa_mu_sim <- kappa_mu_sim_thr_ind_logit(mu, intercept2, beta2, W2, Z2, hyper_fixed)

}
