% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_w1_sim_JM}
\alias{kappa_w1_sim_JM}
\title{Function to simulate the kappa_w1 for joint model}
\usage{
kappa_w1_sim_JM(W1, node1, node2, hyper_fixed)
}
\arguments{
\item{W1}{}

\item{node1}{}

\item{node2}{}
}
\description{
Function to simulate the kappa_w1 for joint model
}
