% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_w2_sim_thr}
\alias{kappa_w2_sim_thr}
\title{Function to simulate the kappa_w2 for threshold model}
\usage{
kappa_w2_sim_thr(W2, node1, node2, hyper_fixed)
}
\arguments{
\item{W2}{}

\item{node1}{}

\item{node2}{}
}
\description{
Function to simulate the kappa_w2 for threshold model
}
