% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_updates.R
\name{latent_MH_update_parallel}
\alias{latent_MH_update_parallel}
\title{Update Latent Parameters Using  Random Walk Metropolis (RWM) in Parallel}
\usage{
latent_MH_update_parallel(
  par_cur,
  par_name,
  loglik_data,
  loglik_latent,
  ns,
  nt,
  var_prop,
  ...
)
}
\arguments{
\item{par_cur}{A vector of length \code{n} containing the current values of the latent parameters.}

\item{par_name}{A character string representing the name of the latent parameter vector being updated.}

\item{loglik_data}{A function representing the log-likelihood at the data level, depending on the latent parameters.}

\item{loglik_latent}{A function representing the log-likelihood at the latent level, often acting as a prior.}

\item{var_prop}{A vector of length \code{n}, representing the variance for the normal random walk proposals.}

\item{...}{Additional arguments to be passed to the log-likelihood functions.}

\item{n}{Integer. The total number of latent parameters.}
}
\value{
A vector of updated latent parameter values, with length \code{n}.
}
\description{
This function updates latent parameters in parallel using the Random Walk Metropolis (RWM) algorithm.
The updates are performed through a Gibbs sampling step with random walk proposals for each latent parameter.
}
\examples{
# Example usage of the latent_MH_update_parallel function:
n <- 20
par_cur <- rnorm(n)
par_name <- "latent_parameter"
var_prop <- rep(0.1, n) # Constant variance for simplicity
loglik_data <- function(latent_parameter, ...) -0.5 * (latent_parameter - 1)^2
loglik_latent <- function(latent_parameter, ...) dnorm(latent_parameter, mean = 1, sd = 0.1, log = TRUE)
updated_params <- latent_MH_update_parallel(par_cur, par_name, loglik_data, loglik_latent, n, var_prop)

}
\seealso{
Other functions involved in the Gibbs sampling process.
}
