% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{lb_ub_markdist}
\alias{lb_ub_markdist}
\title{Lower and Upper Bounds for Landslide Size Distribution Parameters}
\usage{
lb_ub_markdist(mark_dist)
}
\arguments{
\item{mark_dist}{A character string indicating the type of mark distribution. Supported values are:
\itemize{
\item \code{"eGPD"}: Extended generalized Pareto distribution.
\item \code{"bGPD"}: Mixture beta-GPD.
\item \code{"tgGPD"}: Mixture  truncated gamma-GPD.
}}
}
\value{
A list containing two elements:
\itemize{
\item \code{lb}: A numeric vector of the lower bounds for the parameters of the specified mark distribution.
\item \code{ub}: A numeric vector of the upper bounds for the parameters of the specified mark distribution.
}
}
\description{
Provides the lower and upper bounds for the parameters of different mark distributions based on the specified distribution type.
}
\details{
The function returns parameter bounds depending on the type of mark distribution:
\itemize{
\item For \code{"eGPD"}, the lower bounds are \code{c(0, 0)}, and the upper bounds are \code{c(Inf, 1)}.
\item For \code{"bGPD"} and \code{"tgGPD"}, both have lower bounds of \code{c(0)} and upper bounds of \code{c(Inf)}.
}
}
\examples{
# Get parameter bounds for an extended GPD distribution
lb_ub_markdist("eGPD")

# Get parameter bounds for a beta-GPD distribution
lb_ub_markdist("bGPD")

# Get parameter bounds for a truncated GPD distribution
lb_ub_markdist("tgGPD")
}
