% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mark_dist-JM.R
\name{log_lik_GPD_param}
\alias{log_lik_GPD_param}
\title{Log-Likelihood for Generalized Pareto Distribution (GPD) Parameters}
\usage{
log_lik_GPD_param(cur_par, A, threshold)
}
\arguments{
\item{cur_par}{A numeric vector containing the GPD parameters: \code{sigma.GP} (scale parameter) and \code{xi} (shape parameter).}

\item{A}{A numeric vector representing the size data for the GPD model.}

\item{threshold}{A numeric vector representing the threshold values, above which the GPD is considered.}
}
\value{
A numeric value representing the sum of the log-likelihood for all elements of \code{A} that exceed the corresponding \code{threshold}.
}
\description{
This function computes the log-likelihood of a Generalized Pareto Distribution (GPD) for given parameters and data, conditioned on exceeding a specified threshold.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
cur_par <- c(1, 0.2) # sigma.GP = 1, xi = 0.2
A <- rnorm(10, mean = 5)
threshold <- rep(4, 10)
loglik_value <- log_lik_GPD_param(cur_par, A, threshold)
\dontshow{\}) # examplesIf}
}
