% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu_indicator_model.R
\name{log_lik_indicator_model}
\alias{log_lik_indicator_model}
\title{log-likelihood of the mark-distribution process a the data level; i.e., the likelihood of density A | theta_A, mu}
\usage{
log_lik_indicator_model(mu, A)
}
\arguments{
\item{mu}{log-median of the latent size process. More explicitly exp(mu) is the median of the size distribution A}

\item{A}{the size data}

\item{cur_par}{hyper-parameters vector of the continuous mark distributions}

\item{sum_dens}{whether to calculate the log-density individually or the sum of log-likelihood}

\item{ind_zeros_counts}{}
}
\description{
log-likelihood of the mark-distribution process a the data level; i.e., the likelihood of density A | theta_A, mu
}
