% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_latent.R
\name{log_lik_latent_eta_JM}
\alias{log_lik_latent_eta_JM}
\title{Title}
\usage{
log_lik_latent_eta_JM(eta, intercept1, beta1, kappa_eta, W1, Z1)
}
\arguments{
\item{eta}{}

\item{intercept1}{}

\item{beta1}{}

\item{kappa_eta}{}

\item{W1}{}

\item{Z1}{}

\item{Y}{}

\item{A1}{}
}
\description{
Title
}
