% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_latent.R
\name{log_lik_latent_mu_JM}
\alias{log_lik_latent_mu_JM}
\title{Log-Likelihood for Latent \code{mu} in the Joint Model}
\usage{
log_lik_latent_mu_JM(mu, intercept2, W1, W2, beta, beta2, kappa_mu, Z2)
}
\arguments{
\item{mu}{A numeric vector representing the latent variable for which the log-likelihood is being calculated.}

\item{intercept2}{A numeric value representing the intercept of the model.}

\item{W1}{A numeric vector of random effects.}

\item{W2}{A numeric vector of random effects.}

\item{beta}{A numeric value representing sharing parameter used as the coefficient of \code{W1} on \code{mu}.}

\item{beta2}{A numeric vector of coefficients for the covariates in \code{Z2}.}

\item{kappa_mu}{A numeric value representing the precision or scale parameter for \code{mu}.}

\item{Z2}{A matrix of covariates influencing \code{mu}.}
}
\value{
A numeric vector representing the log-likelihood for each value of \code{mu}.
}
\description{
This function calculates the log-likelihood for the latent variable \code{mu} in the context of a joint model.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
mu <- rnorm(10)
intercept2 <- 0.5
W1 <- rnorm(10)
W2 <- rnorm(10)
beta <- 0.3
beta2 <- c(0.3, -0.2)
kappa_mu <- 2
Z2 <- matrix(rnorm(20), nrow = 10, ncol = 2)
loglik_values <- log_lik_latent_mu_JM(mu, intercept2, W1, W2, beta, beta2, kappa_mu, Z2)
\dontshow{\}) # examplesIf}
}
