% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mark_dist-JM.R
\name{log_lik_shape_mixture}
\alias{log_lik_shape_mixture}
\title{log-likelihood of the mark-distribution process at the data level; i.e., the likelihood of density A | theta_A, mu}
\usage{
log_lik_shape_mixture(cur_par, mu, A, mark_dist, threshold, sum_dens = TRUE)
}
\arguments{
\item{cur_par}{hyper-parameters vector of the continuous mark distributions}

\item{mu}{log-median of the latent size process. More explicitly exp(mu) is the median of the size distribution A}

\item{A}{the size data}

\item{mark_dist}{choice of mark_dist of distributions for the size process.}

\item{sum_dens}{whether to calculate the log-density individually or the sum of log-likelihood}
}
\description{
log-likelihood of the mark-distribution process at the data level; i.e., the likelihood of density A | theta_A, mu
}
