% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_threshold-model.R
\name{log_lik_thr}
\alias{log_lik_thr}
\title{log-likelihood of the mark-distribution process a the data level; i.e., the likelihood of density A | theta_A, mu}
\usage{
log_lik_thr(cur_par, mu, A, thr.family, ind_zeros_counts, sum_dens = TRUE)
}
\arguments{
\item{cur_par}{hyper-parameters vector of the continuous mark distributions}

\item{mu}{log-median of the latent size process. More explicitly exp(mu) is the median of the size distribution A}

\item{A}{the size data}

\item{thr.family}{}

\item{ind_zeros_counts}{}

\item{sum_dens}{whether to calculate the log-density individually or the sum of log-likelihood}
}
\description{
log-likelihood of the mark-distribution process a the data level; i.e., the likelihood of density A | theta_A, mu
}
