% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{logistic}
\alias{logistic}
\title{Logistic (Sigmoid) function}
\usage{
logistic(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
A numeric vector of the same length as \code{x}, with values between 0 and 1.
}
\description{
This function computes the logistic function, which transforms the input \code{x} into a value between 0 and 1. The logistic function is widely used to model probabilities and is defined as \code{1 / (1 + exp(-x))}.
}
\details{
The logistic function is used to map real-valued numbers to the (0, 1) interval, making it useful for modeling probabilities in various statistical models, such as binary classification or logistic regression.
}
