% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{pbetat}
\alias{pbetat}
\title{distribution function of beta distribution with support in (0, c)}
\usage{
pbetat(x, u.thr, shape1, shape2, log.p = TRUE)
}
\arguments{
\item{x}{}

\item{shape1}{}

\item{shape2}{}

\item{c}{}

\item{log}{}
}
\description{
distribution function of beta distribution with support in (0, c)
}
