% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{pbGPD}
\alias{pbGPD}
\title{distribution function of mixture of beta in (0, u.thr)  and GPD}
\usage{
pbGPD(x, u.thr, k, xi, sigma, sigma.GP, thr.prob, log.p = TRUE)
}
\arguments{
\item{x}{}

\item{u.thr}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{thr.prob}{}

\item{log}{}
}
\description{
distribution function of mixture of beta in (0, u.thr)  and GPD
}
