% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{pEGPD1}
\alias{pEGPD1}
\title{distribution function of extended GPD of type 1}
\usage{
pEGPD1(x, k, xi, sigma, log = TRUE)
}
\arguments{
\item{x}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{log}{}
}
\description{
distribution function of extended GPD of type 1
}
