% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{pgammat}
\alias{pgammat}
\title{CDF of Truncated Gamma Distribution}
\usage{
pgammat(x, upper.bound, shape, scale, log.p = TRUE)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{upper.bound}{A numeric value indicating the upper bound of the truncation for the Gamma distribution.}

\item{shape}{A positive numeric value representing the shape parameter of the Gamma distribution.}

\item{scale}{A positive numeric value representing the scale parameter of the Gamma distribution.}

\item{log.p}{Logical; if \code{TRUE}, probabilities are given as \code{log(p)}. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of the CDF values of the truncated Gamma distribution at the specified quantiles \code{x}.
}
\description{
This function calculates the cumulative distribution function (CDF) of a Gamma distribution
truncated to an upper bound. It returns either the log of the CDF or the CDF itself, depending
on the input arguments.
}
\details{
The function computes the CDF of a Gamma distribution truncated at the given upper bound.
If \code{log.p = TRUE}, the log of the CDF is returned, otherwise the CDF itself is returned.
The CDF is normalized by the probability that the Gamma variable falls within the truncated range.
}
\examples{
# Example usage of the pgammat function:
x <- seq(0, 5, by = 0.1)
shape <- 2
scale <- 1
upper.bound <- 5
cdf_values <- pgammat(x, upper.bound, shape, scale, log.p = FALSE)

}
\seealso{
\code{\link{pgamma}} for the Gamma distribution function without truncation.
}
