% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rbetat}
\alias{rbetat}
\title{simulate from the beta distribution with support in (0, c)}
\usage{
rbetat(n, u.thr, shape1, shape2)
}
\arguments{
\item{n}{}

\item{shape1}{}

\item{shape2}{}

\item{c}{}
}
\description{
simulate from the beta distribution with support in (0, c)
}
