% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rbGPD}
\alias{rbGPD}
\title{simulating from mixture of beta in (0, u.thr)  and GPD}
\usage{
rbGPD(n, u.thr, k, xi, sigma, sigma.GP, ind)
}
\arguments{
\item{n}{}

\item{u.thr}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{ind}{}
}
\description{
simulating from mixture of beta in (0, u.thr)  and GPD
}
