% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rgammat}
\alias{rgammat}
\title{simulation from truncated Gamma distribution function}
\usage{
rgammat(n, upper.bound, shape, scale = 1)
}
\arguments{
\item{n}{}

\item{shape}{}

\item{scale}{}

\item{range}{}
}
\description{
simulation from truncated Gamma distribution function
}
