% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rMVNormC_eigen}
\alias{rMVNormC_eigen}
\title{Covariance parameterization (Eigen decomposition)}
\usage{
rMVNormC_eigen(n, mu, Sigma)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{p-vector mean}

\item{Sigma}{covariance matrix (p x p)}
}
\value{
matrix of dimension p x n of samples
}
\description{
Covariance parameterization (Eigen decomposition)
}
