% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rMVNormP_eigen}
\alias{rMVNormP_eigen}
\title{Precision Parameterization using Eigen Decomposition}
\usage{
rMVNormP_eigen(n, mu, Sigma)
}
\arguments{
\item{n}{An integer specifying the number of samples to draw.}

\item{mu}{A numeric vector of length \code{p} representing the mean of the distribution.}

\item{Sigma}{A \code{p} x \code{p} precision matrix (the inverse of the covariance matrix).}
}
\value{
A \code{p} x \code{n} matrix of samples drawn from the multivariate normal distribution with mean vector \code{mu} and precision matrix \code{Sigma}.
}
\description{
Generates samples from a multivariate normal distribution with a specified mean vector and precision matrix, utilizing the eigen decomposition for parameterization.
}
\details{
This function uses the eigen decomposition of the precision matrix \code{Sigma} to generate samples from the corresponding multivariate normal distribution. The process is as follows:
\itemize{
\item Perform the truncated eigen decomposition of the precision matrix \code{Sigma}.
\item Draw samples from a standard normal distribution.
\item Transform these samples to have the specified mean \code{mu} and precision structure \code{Sigma}.
}

Note that the precision matrix is the inverse of the covariance matrix. The function inverts the square root transformation to properly adjust for the precision parameterization.
}
\examples{
# Example parameters
mu <- c(0, 0) # Mean vector
Precision <- matrix(c(2, 0.5, 0.5, 1), nrow = 2) # Precision matrix (inverse covariance)
# Generate 100 samples
samples <- rMVNormP_eigen(n = 100, mu = mu, Sigma = Precision
}
