% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rtgGPD}
\alias{rtgGPD}
\title{simulating from mixture of truncated gamma and GPD distribution}
\usage{
rtgGPD(n, u.thr, k, xi, sigma, sigma.GP, ind)
}
\arguments{
\item{n}{}

\item{u.thr}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{ind}{}
}
\description{
simulating from mixture of truncated gamma and GPD distribution
}
