% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{transfo}
\alias{transfo}
\title{Transform parameter to unconstrained scale}
\usage{
transfo(par, lb, ub)
}
\arguments{
\item{par}{scalar parameter}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
Transform parameter to unconstrained scale
}
