% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{trunc_eigen_sqrt}
\alias{trunc_eigen_sqrt}
\title{Truncated Eigen Square Root (or inverse square root)}
\usage{
trunc_eigen_sqrt(Sigma, inv)
}
\arguments{
\item{Sigma}{p x p matrix to compute truncated matrix square root}

\item{inv}{(boolean) whether to compute inverse of square root}
}
\value{
p x k matrix (where k is rank of Sigma)
}
\description{
Designed for Covaraice or Precision matricies - throws error if
there is any large negative Eigenvalues.
}
