% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu_indicator_model.R
\name{update_mu}
\alias{update_mu}
\title{Title}
\usage{
update_mu(mu_mean, kappa.mu, ind_zero, ind.NA, CV)
}
\arguments{
\item{mu_mean}{}

\item{kappa.mu}{}

\item{ind_zero}{}

\item{ind.NA}{}
}
\description{
Title
}
