% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ICAR_RE.R
\name{W2_sim_Gibbs_componentwise_thr}
\alias{W2_sim_Gibbs_componentwise_thr}
\title{Simulate iCAR Random Effects \code{W2} Using Gibbs Sampling for Threshold Model}
\usage{
W2_sim_Gibbs_componentwise_thr(
  node_index,
  W2,
  mean_mu_latent,
  kappa_w2,
  kappa_mu,
  nbd_info,
  no_of_nbd
)
}
\arguments{
\item{node_index}{An integer index indicating the specific node for which \code{W2} is being simulated.}

\item{W2}{A numeric vector representing the current values of the latent variable \code{W2}.}

\item{mean_mu_latent}{A numeric value representing the mean of the latent variable \code{mu}.}

\item{kappa_w2}{A numeric value representing the precision parameter associated with \code{W2}.}

\item{kappa_mu}{A numeric value representing the precision parameter associated with \code{mu}.}

\item{nbd_info}{A list containing neighborhood information for each node, indicating which nodes are neighbors.}

\item{no_of_nbd}{A numeric vector representing the number of neighbors for each node.}
}
\value{
A numeric value representing the simulated value for \code{W2} at the specified \code{node_index}.
}
\description{
This function simulates the iCAR random effects \code{W2} for the threshold model using a component-wise Gibbs sampling approach. It updates \code{W2} based on the local neighborhood information and the latent mean.
}
\examples{
# Example data for the threshold model
W2 <- rnorm(100)
mean_mu_latent <- 0.5
kappa_w2 <- 1.0
kappa_mu <- 2.0
nbd_info <- list(1 = c(2, 3), 2 = c(1, 3), 3 = c(1, 2))
no_of_nbd <- c(2, 2, 2)

# Simulate W2 for a specific node index
simulated_W2 <- W2_sim_Gibbs_componentwise_thr(1, W2, mean_mu_latent, kappa_w2, kappa_mu, nbd_info, no_of_nbd)
}
