% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta2_sim_JM}
\alias{beta2_sim_JM}
\title{Function to simulate beta2 for joint model}
\usage{
beta2_sim_JM(mu, intercept2, beta, kappa_mu, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{intercept2}{}

\item{beta}{}

\item{kappa_mu}{}

\item{W1}{}

\item{W2}{}

\item{Z2}{}
}
\description{
Function to simulate beta2 for joint model
}
