% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{beta_sim_JM}
\alias{beta_sim_JM}
\title{Simulate Sharing Coefficient \code{beta} of Joint Model}
\usage{
beta_sim_JM(mu, intercept2, kappa_mu, beta2, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the response variable.}

\item{intercept2}{A numeric value representing the intercept term in the model.}

\item{kappa_mu}{A numeric value representing the precision parameter associated with \code{mu}.}

\item{beta2}{A numeric vector of coefficients for covariates in \code{Z2}.}

\item{W1}{A numeric vector representing the first set of latent random effects in the model.}

\item{W2}{A numeric vector representing the second set of latent random effects in the model.}

\item{Z2}{A numeric matrix of covariates corresponding to the response variable.}

\item{hyper_fixed}{A list containing hyperparameters for the Normal prior distribution, specifically the precision for \code{beta}.}
}
\value{
A simulated value for \code{beta} from its full conditional normal distribution.
}
\description{
This function simulates the sharing coefficient \code{beta} for a joint model using its full conditional distribution, which is normally distributed with precision based on covariates and latent variables.
}
\examples{
# Example data for the joint model
mu <- rnorm(100)
intercept2 <- 0.5
kappa_mu <- 2
beta2 <- rnorm(5)
W1 <- rnorm(100)
W2 <- rnorm(100)
Z2 <- matrix(rnorm(500), 100, 5)
hyper_fixed <- list(beta = 1.5)

# Simulate beta for joint model
beta_sim <- beta_sim_JM(mu, intercept2, kappa_mu, beta2, W1, W2, Z2, hyper_fixed)
}
