% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{cond.pbGPD}
\alias{cond.pbGPD}
\title{Title}
\usage{
cond.pbGPD(x, mu, u.thr, k, xi, sigma, sigma.GP, thr.prob)
}
\arguments{
\item{x}{x is high quantile, greater than u.thr}

\item{mu}{}

\item{u.thr}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{thr.prob}{}

\item{log.p}{}
}
\description{
Title
}
