% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dbetat}
\alias{dbetat}
\title{Density Function of Scaled Beta Distribution}
\usage{
dbetat(x, u.thr, shape1, shape2, log = TRUE)
}
\arguments{
\item{x}{A numeric vector of quantiles for which the density is to be evaluated.}

\item{u.thr}{A positive numeric value specifying the upper bound of the support of the Beta distribution.}

\item{shape1}{A positive numeric value representing the first shape parameter of the Beta distribution.}

\item{shape2}{A positive numeric value representing the second shape parameter of the Beta distribution.}

\item{log}{Logical; if \code{TRUE}, the log-density is returned. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of density values corresponding to the quantiles in \code{x}.
}
\description{
This function computes the density of a Beta distribution scaled to have its support in the interval \eqn{(0, u.thr)}.
}
\details{
The standard Beta distribution has support on the interval \eqn{(0, 1)}. This function rescales the distribution so that
its support lies in \eqn{(0, u.thr)}. The density is computed by applying a change of variables to the standard Beta
density function.

If \code{log} is \code{TRUE}, the log of the density is returned.
}
\examples{
# Evaluate the density of the scaled Beta distribution
x <- seq(0.01, 5, length.out = 100)
u.thr <- 5
shape1 <- 2
shape2 <- 3
density_values <- dbetat(x, u.thr, shape1, shape2, log = FALSE)
plot(x, density_values, type = "l", main = "Scaled Beta Density", ylab = "Density", xlab = "x")

}
\seealso{
\code{\link{dbeta}} for the standard Beta distribution density.
}
