% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dEGPD1}
\alias{dEGPD1}
\title{density function of extended GPD of type 1}
\usage{
dEGPD1(x, k, xi, sigma, log = TRUE)
}
\arguments{
\item{x}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{log}{}
}
\description{
density function of extended GPD of type 1
}
