% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dgammat}
\alias{dgammat}
\title{density of truncated gamma distribution}
\usage{
dgammat(x, upper.bound, shape, scale, log = TRUE)
}
\arguments{
\item{x}{}

\item{shape}{}

\item{scale}{}

\item{log}{}

\item{range}{}
}
\description{
density of truncated gamma distribution
}
