% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dlogjac_inv_transfo}
\alias{dlogjac_inv_transfo}
\title{Gradient of the Jacobean}
\usage{
dlogjac_inv_transfo(tpar, lb, ub)
}
\arguments{
\item{tpar}{the transformed parameter}

\item{lb}{lower bound of parameter}

\item{ub}{upper bound of the parameter}
}
\description{
Gradient of the Jacobean
}
