% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{dtgGPD}
\alias{dtgGPD}
\title{density function of mixture of truncated gamma and GPD distribution}
\usage{
dtgGPD(x, u.thr, k, xi, sigma, sigma.GP, thr.prob, log = TRUE)
}
\arguments{
\item{x}{}

\item{k}{}

\item{xi}{}

\item{sigma}{}

\item{thr.prob}{}

\item{log}{}

\item{range}{}
}
\description{
density function of mixture of truncated gamma and GPD distribution
}
