% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_GPD_param}
\alias{gamma_prior_GPD_param}
\title{Prior Distributions for GPD Parameters}
\usage{
gamma_prior_GPD_param(cur_par, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{A numeric vector containing the current parameter values of the GPD: \code{cur_par[1]} is the scale parameter (\code{sigma}) and \code{cur_par[2]} is the shape parameter (\code{xi}).}

\item{hyper.mu_fixed}{A numeric vector of length 2 providing the shape and rate hyperparameters of the gamma prior distribution for the \code{sigma.GP} parameter.}
}
\value{
A numeric value representing the log-prior density of the GPD parameters.
}
\description{
Computes the log-prior density for the parameters of a generalized Pareto distribution (GPD), specifically for the scale (\code{sigma}) and shape (\code{xi}) parameters.
}
\details{
The function computes the log-prior density for:
\itemize{
\item \strong{\code{sigma.GP}}: Assumed to have a gamma prior with shape and rate hyperparameters specified in \code{hyper.mu_fixed}.
\item \strong{\code{xi}}: Assumed to have a gamma prior with shape 1 and rate 15.
}
}
\examples{
# Example for calculating the log-prior density for GPD parameters
cur_par <- c(2, 0.5) # sigma.GP and xi
hyper.mu_fixed <- c(1, 0.5) # shape and rate hyperparameters for sigma.GP
gamma_prior_GPD_param(cur_par, hyper.mu_fixed)
}
