% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_markdist_JM}
\alias{gamma_prior_markdist_JM}
\title{Prior Distributions for Joint Model's Landslide Size Distributions}
\usage{
gamma_prior_markdist_JM(cur_par, mark_dist, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{A numeric vector representing the current parameter values of the mark distribution. For eGPD, includes both \code{k} and \code{xi}. For bGPD and tgGPD, only \code{k} is considered.}

\item{mark_dist}{A character string indicating the type of mark distribution. Possible values are \code{"eGPD"}, \code{"bGPD"}, and \code{"tgGPD"}.}

\item{hyper.mu_fixed}{A numeric vector of length 2, providing the shape and rate hyperparameters of the gamma prior distribution for the mark distribution parameter.}
}
\value{
A numeric value representing the log-prior density of the parameters under the specified mark distribution family.
}
\description{
Computes the log-prior density for the parameters of the landslide size distribution in a joint model, supporting various distribution families like extended GPD (eGPD), beta-GPD (bGPD), and truncated gamma-GPD (tgGPD).
}
\details{
This function calculates the log-prior density for the parameters of the mark distribution based on the specified family:
\itemize{
\item \strong{eGPD}: The prior for \code{k} is a gamma distribution, and the prior for \code{xi} is also a gamma distribution with shape 1 and rate 15.
\item \strong{bGPD}: Only the shape parameter \code{k} has a gamma prior distribution with the specified hyperparameters.
\item \strong{tgGPD}: Similar to bGPD, only \code{k} has a gamma prior distribution.
}
}
\examples{
# Prior for extended GPD (eGPD)
cur_par <- c(2, 0.5) # k and xi
mark_dist <- "eGPD"
hyper.mu_fixed <- c(1, 0.5)
gamma_prior_markdist_JM(cur_par, mark_dist, hyper.mu_fixed)

# Prior for beta-GPD (bGPD)
cur_par <- c(3)
mark_dist <- "bGPD"
hyper.mu_fixed <- c(1, 0.5)
gamma_prior_markdist_JM(cur_par, mark_dist, hyper.mu_fixed)

# Prior for truncated gamma-GPD (tgGPD)
cur_par <- c(4)
mark_dist <- "tgGPD"
hyper.mu_fixed <- c(2, 1)
gamma_prior_markdist_JM(cur_par, mark_dist, hyper.mu_fixed)
}
