% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hyperparameters.R
\name{gamma_prior_markdist_thr}
\alias{gamma_prior_markdist_thr}
\title{prior distributions for the mark distributions}
\usage{
gamma_prior_markdist_thr(cur_par, thr.family, hyper.mu_fixed)
}
\arguments{
\item{cur_par}{}

\item{hyper.mu_fixed}{}

\item{mark_dist}{}
}
\description{
prior distributions for the mark distributions
}
