% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-misisng locations.R
\name{impute.NA.A}
\alias{impute.NA.A}
\title{Imputing NA for landslides sizes in case of with sample (WS)}
\usage{
impute.NA.A(
  CV,
  ind_NA_A,
  ind_zeros_counts,
  mu,
  thr.prob,
  cur_par,
  mark_dist,
  threshold,
  thr.exceed.ind = NULL
)
}
\arguments{
\item{ind_zeros_counts}{}

\item{mu}{}

\item{thr.prob}{}

\item{cur_par}{}

\item{mark_dist}{}

\item{threshold}{}

\item{thr.exceed.ind}{}
}
\description{
Imputing NA for landslides sizes in case of with sample (WS)
}
