% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-missing locations.R
\name{impute.NA.Y}
\alias{impute.NA.Y}
\title{Impute NA Values for Landslide Counts}
\usage{
impute.NA.Y(ind_NA_Y, eta, CV)
}
\arguments{
\item{ind_NA_Y}{A logical vector indicating which elements in the count data are missing (NA).}

\item{eta}{A numeric vector of linear predictors for the Poisson distribution.}

\item{CV}{A character string specifying the cross-validation method to use. Accepts either \code{"OOS"} for out-of-sample
cross-validation or \code{"WS"} for within-sample cross-validation.}
}
\value{
A vector of imputed values, replacing NA elements in the original count data.
If \code{CV = "OOS"}, only the NA values are imputed. If \code{CV = "WS"}, all values are replaced with new counts.
}
\description{
This function imputes missing (NA) values in landslide count data using a Poisson distribution.
}
\examples{
ind_NA_Y <- c(TRUE, FALSE, TRUE, FALSE)
eta <- c(1.2, 0.5, 1.8, 0.3)
CV <- "OOS"
imputed_values <- impute.NA.Y(ind_NA_Y, eta, CV)

}
