% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_all_other_param_JM}
\alias{init_fun_all_other_param_JM}
\title{nitial values of all the other parameters, that does not depends on the choice of mark distributions}
\usage{
init_fun_all_other_param_JM(
  Z1,
  Z2,
  A,
  Y,
  seed,
  simulation,
  threshold,
  mark_dist,
  model_type,
  thr.acces.ind
)
}
\arguments{
\item{Z1}{}

\item{Z2}{}

\item{A}{}

\item{Y}{}

\item{seed}{}

\item{simulation}{}

\item{threshold}{}

\item{thr.acces.ind}{}
}
\description{
nitial values of all the other parameters, that does not depends on the choice of mark distributions
}
