% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_hyper.GP_JM}
\alias{init_fun_hyper.GP_JM}
\title{Initial values and parameter names for hyperparameter for GP}
\usage{
init_fun_hyper.GP_JM(mark_dist, seed)
}
\arguments{
\item{mark_dist}{}

\item{seed}{}
}
\description{
Initial values and parameter names for hyperparameter for GP
}
