% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_values.R
\name{init_fun_hyper.mu_JM}
\alias{init_fun_hyper.mu_JM}
\title{Initialize Parameters for Hyperparameter in the Bulk}
\usage{
init_fun_hyper.mu_JM(mark_dist, seed)
}
\arguments{
\item{mark_dist}{A character string indicating the type of marginal distribution. Possible values are:
\code{"eGPD"} for extended Generalized Pareto Distribution, \code{"bGPD"} for a mixture of beta-GPD, or \code{"tgGPD"} for a mixture of truncated gamma-GPD.}

\item{seed}{A numeric value to set the seed for random number generation to ensure reproducibility.}
}
\value{
A list containing:
\item{log.hyper.mu.init}{A named list of initial values for the hyperparameters, such as \code{k} and \code{xi}.}
\item{log.hyper.mu.name}{An expression vector with the names of the parameters for better display in plots or tables.}
}
\description{
This function initializes the values and names for the hyperparameters based on the specified marginal distribution.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
mark_dist <- "eGPD"
seed <- 123
init_values <- init_fun_hyper.mu_JM(mark_dist, seed)
\dontshow{\}) # examplesIf}
}
