% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept1_sim_JM}
\alias{intercept1_sim_JM}
\title{Function to simulate the intercept1 for joint model}
\usage{
intercept1_sim_JM(eta, beta1, W1, kappa_eta, Z1, hyper_fixed)
}
\arguments{
\item{eta}{}

\item{beta1}{}

\item{W1}{}

\item{kappa_eta}{}

\item{Z1}{}
}
\description{
Function to simulate the intercept1 for joint model
}
