% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept2_sim_JM}
\alias{intercept2_sim_JM}
\title{Simulate Intercept \code{intercept2} of \code{mu} for Joint Model}
\usage{
intercept2_sim_JM(mu, beta2, beta, W1, W2, kappa_mu, Z2, hyper_fixed)
}
\arguments{
\item{mu}{A numeric vector representing the response variable for the second part of the model.}

\item{beta2}{A numeric vector of coefficients for the covariates in \code{Z2}.}

\item{beta}{A numeric value representing the effect of the latent variable \code{W1} on \code{mu}.}

\item{W1}{A numeric vector representing the latent random effects for the first part of the model.}

\item{W2}{A numeric vector representing the latent random effects for the second part of the model.}

\item{kappa_mu}{A numeric value representing the precision parameter associated with \code{mu}.}

\item{Z2}{A numeric matrix of covariates corresponding to the response variable \code{mu}.}

\item{hyper_fixed}{A list containing hyperparameters, specifically the prior precision for \code{intercept2}.}
}
\value{
A simulated value for \code{intercept2} from its full conditional normal distribution.
}
\description{
This function simulates the intercept parameter \code{intercept2} for a joint model based on its full conditional distribution, which is normally distributed with precision derived from the latent variables, covariates, and random effects.
}
\examples{
# Example data for the joint model
mu <- rnorm(100)
beta2 <- rnorm(5)
beta <- 0.5
W1 <- rnorm(100)
W2 <- rnorm(100)
kappa_mu <- 2
Z2 <- matrix(rnorm(500), 100, 5)
hyper_fixed <- list(intercept1 = 1.5)

# Simulate intercept2 for joint model
intercept2_sim <- intercept2_sim_JM(mu, beta2, beta, W1, W2, kappa_mu, Z2, hyper_fixed)
}
