% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{intercept2_sim_thr}
\alias{intercept2_sim_thr}
\title{Function to simulate the intercept2 for threshold model}
\usage{
intercept2_sim_thr(mu, beta2, beta, W2, kappa_mu, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{beta2}{}

\item{beta}{}

\item{W2}{}

\item{kappa_mu}{}

\item{Z2}{}
}
\description{
Function to simulate the intercept2 for threshold model
}
