% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{inv_transfo}
\alias{inv_transfo}
\title{Transform the parameters back to original scales}
\usage{
inv_transfo(tpar, lb, ub)
}
\arguments{
\item{tpar}{underlying parameter}

\item{lb}{lower support of the parameters}

\item{ub}{upper support of the parameters}
}
\value{
transformed parameter
}
\description{
Transform the parameters back to original scales
}
