% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_eta_sim_JM}
\alias{kappa_eta_sim_JM}
\title{Simulate Precision Parameter \code{kappa_eta} of \code{eta} for Joint Model}
\usage{
kappa_eta_sim_JM(eta, intercept1, beta1, W1, Z1, hyper_fixed)
}
\arguments{
\item{eta}{A numeric vector representing the latent variable for the model.}

\item{intercept1}{A numeric value representing the intercept term in the joint model.}

\item{beta1}{A numeric vector of regression coefficients associated with the covariates \code{Z1}.}

\item{W1}{A numeric vector representing the latent random effects for the joint model.}

\item{Z1}{A numeric matrix of covariates associated with \code{eta}.}

\item{hyper_fixed}{A list containing hyperparameters for the Gamma prior distribution, specifically the shape and rate parameters for \code{kappa_eta}.}
}
\value{
A simulated value for \code{kappa_eta} from the Gamma distribution.
}
\description{
This function simulates the precision parameter \code{kappa_eta} for a joint model from a Gamma full conditional distribution.
}
\examples{
# Sample latent variable and covariates
eta <- rnorm(100)
intercept1 <- 0.5
beta1 <- rnorm(10)
W1 <- rnorm(100)
Z1 <- matrix(rnorm(1000), ncol = 10)
hyper_fixed <- list(kappa_eta = c(shape = 2, rate = 1))

# Simulate kappa_eta for joint model
kappa_sim <- kappa_eta_sim_JM(eta, intercept1, beta1, W1, Z1, hyper_fixed)
}
