% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_mu_sim_JM}
\alias{kappa_mu_sim_JM}
\title{Function to simulate the kappa_mu for joint model}
\usage{
kappa_mu_sim_JM(mu, intercept2, beta2, beta, W1, W2, Z2, hyper_fixed)
}
\arguments{
\item{mu}{}

\item{intercept2}{}

\item{beta2}{}

\item{beta}{}

\item{W1}{}

\item{W2}{}

\item{Z2}{}
}
\description{
Function to simulate the kappa_mu for joint model
}
