% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_w2_sim_thr}
\alias{kappa_w2_sim_thr}
\title{Simulate the Precision Parameter of iCAR Random Effects for Threshold Model}
\usage{
kappa_w2_sim_thr(W2, node1, node2, hyper_fixed)
}
\arguments{
\item{W2}{A numeric vector representing the latent variable (iCAR random effects) for the model.}

\item{node1}{An integer vector indicating the first set of connected nodes (indices) involved in the computation.}

\item{node2}{An integer vector indicating the second set of connected nodes (indices) involved in the computation.}

\item{hyper_fixed}{A list containing hyperparameters for the Gamma prior distribution of \code{kappa_w2}, specifically the shape and rate parameters (\code{kappa_w2[1]} for shape and \code{kappa_w2[2]} for rate).}
}
\value{
A simulated value for \code{kappa_w2} from the Gamma distribution.
}
\description{
This function simulates the parameter \code{kappa_w2}, the precision parameter of the intrinsic Conditional Autoregressive (iCAR) model for a threshold model, from a full conditional Gamma distribution.
}
\examples{
# Sample latent variable
W2 <- rnorm(100)
node1 <- sample(1:100, 50, replace = TRUE)
node2 <- sample(1:100, 50, replace = TRUE)
hyper_fixed <- list(kappa_w2 = c(shape = 2, rate = 1))

# Simulate kappa_w2 for threshold model
kappa_sim <- kappa_w2_sim_thr(W2, node1, node2, hyper_fixed)

}
