% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_coeff.R
\name{kappa_w2_sim_thr_ind}
\alias{kappa_w2_sim_thr_ind}
\title{Simulate the Precision Parameter \code{kappa_w2} for Threshold Indicator Models}
\usage{
kappa_w2_sim_thr_ind(W2, node1, node2, hyper_fixed)
}
\arguments{
\item{W2}{A numeric vector representing the latent variable (iCAR) random effects for the model.}

\item{node1}{A numeric vector indicating the first set of nodes (indices) involved in the computation.}

\item{node2}{A numeric vector indicating the second set of nodes (indices) involved in the computation.}

\item{hyper_fixed}{A list containing hyperparameters for the Gamma prior distribution, specifically the shape and rate parameters for \code{kappa_w2}.}
}
\value{
A simulated value for \code{kappa_w2} from the Gamma distribution.
}
\description{
This function simulates the precision parameter \code{kappa_w2} for threshold indicator models based on a full conditional Gamma distribution.
}
\examples{
# Sample latent variable
W2 <- rnorm(100)
node1 <- sample(1:100, 50, replace = TRUE)
node2 <- sample(1:100, 50, replace = TRUE)
hyper_fixed <- list(kappa_w2 = c(shape = 2, rate = 1))

# Simulate kappa_w2 for threshold indicator model
kappa_sim <- kappa_w2_sim_thr_ind(W2, node1, node2, hyper_fixed)

}
