% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_updates.R
\name{latent_MH_update_parallel_constrain}
\alias{latent_MH_update_parallel_constrain}
\title{Update the latent parameters in parallel (matrix forms) using the Random Walk Metropolis (RWM) in parallel}
\usage{
latent_MH_update_parallel_constrain(
  par_cur,
  par_name,
  loglik_data,
  loglik_latent,
  ns,
  nt,
  var_prop,
  transform,
  lb,
  ub,
  ...
)
}
\arguments{
\item{par_cur}{A matrix of dimension ns x nt containing the current parameter values.}

\item{par_name}{Name of the latent parameter vectors being updated here.}

\item{loglik_data}{Log-likelihood at the data level, which depends on the underlying
latent parameters.}

\item{loglik_latent}{Log-likelihood at the latent level (may be considered as a prior).}

\item{ns}{Number of spatial locations.}

\item{nt}{Number of time points.}

\item{var_prop}{Tuning parameter (variance) of the normal random walk proposals.}

\item{...}{Additional arguments to be passed to the log-likelihood functions.}
}
\value{
A list containing the updated parameter matrix and the number of times a
parameter is accepted.
}
\description{
This function performs updates for latent parameters in a parallel manner using the
Random Walk Metropolis (RWM) algorithm. It updates the latent parameters based on
random walk proposals in a Gibbs sampling step to achieve convergence.
}
\examples{
# Example usage of the latent_MH_update_parallel function:
par_cur <- matrix(rnorm(5 * 4), nrow = 5, ncol = 4)
par_name <- "latent_parameter"
ns <- 5
nt <- 4
var_prop <- 0.1
acc_par <- matrix(runif(ns * nt), nrow = ns, ncol = nt)
loglik_data <- function(latent_parameter) -0.5 * (latent_parameter - 1)^2
loglik_latent <- function(latent_parameter) dnorm(latent_parameter, mean = 1, sd = 0.1, log = TRUE)
updated_params <- latent_MH_update_parallel(par_cur, par_name, loglik_data, loglik_latent, ns, nt, var_prop, acc_par)

}
\seealso{
Other functions used in the Gibbs sampling process.
}
