% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_latent.R
\name{log_lik_eta_JM}
\alias{log_lik_eta_JM}
\title{Title}
\usage{
log_lik_eta_JM(Y, eta)
}
\arguments{
\item{Y}{}

\item{eta}{}

\item{intercept1}{}

\item{beta1}{}

\item{kappa_eta}{}

\item{W1}{}

\item{Z1}{}

\item{A1}{}
}
\description{
Title
}
