% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mark_dist-JM.R
\name{log_lik_GPD_param}
\alias{log_lik_GPD_param}
\title{Title}
\usage{
log_lik_GPD_param(cur_par, A, threshold)
}
\arguments{
\item{cur_par}{}

\item{A}{}

\item{threshold}{}

\item{thr.exceed.ind}{}

\item{thr.prob}{}

\item{mu}{}

\item{mark_dist}{}

\item{ind_zeros_counts}{}

\item{sum_dens}{}
}
\description{
Title
}
