% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_latent.R
\name{log_lik_latent_eta_JM}
\alias{log_lik_latent_eta_JM}
\title{Log-Likelihood for Latent \code{eta} in Joint Model}
\usage{
log_lik_latent_eta_JM(eta, intercept1, beta1, kappa_eta, W1, Z1)
}
\arguments{
\item{eta}{A numeric vector representing the latent variable for which the log-likelihood is being calculated.}

\item{intercept1}{A numeric value representing the intercept for the first part of the model.}

\item{beta1}{A numeric vector of coefficients for the covariates in \code{Z1}.}

\item{kappa_eta}{A numeric value representing the precision or scale parameter for \code{eta}.}

\item{W1}{A numeric vector of random effects or covariates influencing \code{eta}.}

\item{Z1}{A matrix of covariates influencing \code{eta}.}
}
\value{
A numeric vector representing the log-likelihood for each value of \code{eta}.
}
\description{
This function calculates the log-likelihood for the latent variable \code{eta} in the context of a joint model.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
eta <- rnorm(10)
intercept1 <- 0.5
W1 <- rnorm(10)
beta1 <- c(0.3, -0.2)
kappa_eta <- 2
Z1 <- matrix(rnorm(20), nrow = 10, ncol = 2)
loglik_values <- log_lik_latent_eta_JM(eta, intercept1, beta1, kappa_eta, W1, Z1)
\dontshow{\}) # examplesIf}
}
