% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_latent.R
\name{log_lik_latent_mu_JM}
\alias{log_lik_latent_mu_JM}
\title{Title}
\usage{
log_lik_latent_mu_JM(mu, intercept2, W1, W2, beta, beta2, kappa_mu, Z2)
}
\arguments{
\item{mu}{}

\item{intercept2}{}

\item{W1}{}

\item{W2}{}

\item{beta}{}

\item{beta2}{}

\item{kappa_mu}{}

\item{Z2}{}

\item{A}{}

\item{log.hyper.mu}{}

\item{family}{}
}
\description{
Title
}
