% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mark_dist-JM.R
\name{log_lik_shape_mixture}
\alias{log_lik_shape_mixture}
\title{Log-liikelihood of the use mark-distribution}
\usage{
log_lik_shape_mixture(cur_par, mu, A, mark_dist, threshold, sum_dens = TRUE)
}
\arguments{
\item{cur_par}{A numeric vector representing the hyperparameters of the continuous mark distributions. For example, \code{k} and \code{xi} in the case of eGPD.}

\item{mu}{A numeric vector representing the log-median of the latent size process. Explicitly, \code{exp(mu)} is the median of the size distribution \code{A}.}

\item{A}{A numeric vector representing the landslide size data.}

\item{mark_dist}{A character string indicating the choice of mark distribution for the size process. Can be \code{"eGPD"}, \code{"bGPD"}, or \code{"tgGPD"}.}

\item{threshold}{A numeric vector representing the threshold values for the size data \code{A}.}

\item{sum_dens}{A logical value indicating whether to calculate the sum of the log-likelihood (\code{TRUE}) or return the log-density values individually (\code{FALSE}).}
}
\value{
A numeric value representing the sum of log-likelihood values if \code{sum_dens = TRUE}, or a numeric vector of log-density values if \code{sum_dens = FALSE}.
}
\description{
This function computes the log-likelihood of the size distribution  given the parameters, latent size process, and data. It supports three types of mark distributions: extended Generalized Pareto Distribution (eGPD), mixture of beta-GPD (bGPD), and mixture of truncated gamma-GPD (tgGPD).
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage (not meant to be run directly):
cur_par <- c(2, 0.1) # example parameters for eGPD
mu <- rnorm(10)
A <- rgamma(10, shape = 2, scale = 1)
mark_dist <- "eGPD"
threshold <- rep(1, 10)
sum_dens <- TRUE
loglik_value <- log_lik_shape_mixture(cur_par, mu, A, mark_dist, threshold, sum_dens)
\dontshow{\}) # examplesIf}
}
