% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{pgammat}
\alias{pgammat}
\title{cdf of truncated Gamma distribution function}
\usage{
pgammat(x, upper.bound, shape, scale, log.p = TRUE)
}
\arguments{
\item{x}{}

\item{shape}{}

\item{scale}{}

\item{log.p}{}

\item{range}{}
}
\description{
cdf of truncated Gamma distribution function
}
