% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rbetat}
\alias{rbetat}
\title{Simulate from Scaled Beta Distribution}
\usage{
rbetat(n, u.thr, shape1, shape2)
}
\arguments{
\item{n}{An integer specifying the number of random variates to generate.}

\item{u.thr}{A positive numeric value specifying the upper bound of the support of the Beta distribution.}

\item{shape1}{A positive numeric value representing the first shape parameter of the Beta distribution.}

\item{shape2}{A positive numeric value representing the second shape parameter of the Beta distribution.}
}
\value{
A numeric vector of length \code{n} containing random variates from the scaled Beta distribution.
}
\description{
This function generates random variates from a Beta distribution scaled to have its support in the interval \eqn{(0, u.thr)}.
}
\details{
The standard Beta distribution has support on the interval \eqn{(0, 1)}. This function rescales the distribution
so that its support lies in \eqn{(0, u.thr)}. The random variates are generated by sampling from a standard Beta
distribution and then scaling the results by \code{u.thr}.
}
\examples{
# Generate 100 random variates from a scaled Beta distribution
set.seed(123)
n <- 100
u.thr <- 5
shape1 <- 2
shape2 <- 3
random_variates <- rbetat(n, u.thr, shape1, shape2)
hist(random_variates, main = "Scaled Beta Distribution", xlab = "Value", breaks = 20)

}
\seealso{
\code{\link{rbeta}} for the standard Beta distribution random variate generation.
}
