% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rbGPD}
\alias{rbGPD}
\title{Simulate from Mixture of Beta and GPD Distributions}
\usage{
rbGPD(n, u.thr, k, xi, sigma, sigma.GP, ind)
}
\arguments{
\item{n}{An integer specifying the number of observations to generate.}

\item{u.thr}{A positive numeric value indicating the threshold separating the truncated beta and GPD components.}

\item{k}{A positive numeric value representing the shape1 parameter of the truncated beta distribution.}

\item{xi}{A numeric value representing the shape parameter of the GPD.}

\item{sigma}{A numeric vector of scale parameters for the truncated beta distribution, with length equal to \code{n}.}

\item{sigma.GP}{A positive numeric value representing the scale parameter of the GPD.}

\item{ind}{A logical vector of length \code{n}, indicating which observations should be generated from the GPD component (\code{TRUE}) and which from the truncated beta component (\code{FALSE}).}
}
\value{
A numeric vector of length \code{n} containing simulated values from the mixture distribution.
}
\description{
This function simulates values from a mixture distribution, composed of a beta distribution truncated to (0, \code{u.thr}) and a Generalized Pareto Distribution (GPD) for values greater than \code{u.thr}.
}
\details{
This function generates values from a mixture of two distributions:
\itemize{
\item The truncated beta distribution is used for values below the threshold \code{u.thr}.
\item The GPD is used for values greater than \code{u.thr}.
The argument \code{ind} is used to determine which distribution to sample from for each observation.
}
}
\examples{
# Example usage of rbGPD function
n <- 100
u.thr <- 5
k <- 2
xi <- 0.5
sigma <- rep(1, n)
sigma.GP <- 1.5
ind <- runif(n) < 0.3  # 30\% of the samples will be from GPD
simulated_values <- rbGPD(n, u.thr, k, xi, sigma, sigma.GP, ind)
hist(simulated_values, breaks = 30, main = "Simulated Mixture of Beta and GPD")
}
\seealso{
\code{\link{rbetat}} for simulating from the truncated beta distribution,
\code{\link[evd]{rgpd}} for simulating from the Generalized Pareto Distribution.
}
