% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rMVNormC_eigen}
\alias{rMVNormC_eigen}
\title{Covariance Parameterization using Eigen Decomposition}
\usage{
rMVNormC_eigen(n, mu, Sigma)
}
\arguments{
\item{n}{An integer specifying the number of samples to draw.}

\item{mu}{A numeric vector of length \code{p} representing the mean of the distribution.}

\item{Sigma}{A \code{p} x \code{p} covariance matrix.}
}
\value{
A \code{p} x \code{n} matrix of samples drawn from the multivariate normal distribution with mean vector \code{mu} and covariance matrix \code{Sigma}.
}
\description{
Generates samples from a multivariate normal distribution with a specified mean vector and covariance matrix, utilizing the eigen decomposition for parameterization.
}
\details{
This function uses the eigen decomposition of the covariance matrix \code{Sigma} to generate samples from the corresponding multivariate normal distribution. The process is as follows:
\itemize{
\item Perform the truncated eigen decomposition of \code{Sigma}.
\item Draw samples from a standard normal distribution.
\item Transform these samples to have the specified mean \code{mu} and covariance structure \code{Sigma}.
}
}
\examples{
# Example parameters
mu <- c(0, 0) # Mean vector
Sigma <- matrix(c(2, 0.5, 0.5, 1), nrow = 2) # Covariance matrix
# Generate 100 samples
samples <- rMVNormC_eigen(n = 100, mu = mu, Sigma = Sigma)
}
