% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rMVNormP_eigen}
\alias{rMVNormP_eigen}
\title{Precision parameterization (Eigen decomposition)}
\usage{
rMVNormP_eigen(n, mu, Sigma)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{p-vector mean}

\item{Sigma}{precision matrix (p x p)}
}
\value{
matrix of dimension p x n of samples
}
\description{
Precision parameterization (Eigen decomposition)
}
