% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-supporting-functions.R
\name{rtnorm}
\alias{rtnorm}
\title{Univariate truncated normal sampler}
\usage{
rtnorm(n = 1, mean = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{n}{\link{integer} sample size}

\item{mean}{\link{numeric} scalar location parameter}

\item{sd}{\link{numeric} scalar scale parameter}

\item{a}{\link{numeric} lower bound}

\item{b}{\link{numeric} upper bound}
}
\description{
This implementations works for a < 37
or b > 37
}
