% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ICAR_RE.R
\name{W1_sim_Gibbs_componentwise_JM}
\alias{W1_sim_Gibbs_componentwise_JM}
\title{Title}
\usage{
W1_sim_Gibbs_componentwise_JM(
  node_index,
  W1,
  mean_eta_latent,
  mean_mu_latent,
  kappa_w1,
  kappa_eta,
  kappa_mu,
  beta,
  nbd_info,
  no_of_nbd
)
}
\arguments{
\item{node_index}{}

\item{W1}{}

\item{mean_eta_latent}{}

\item{mean_mu_latent}{}

\item{kappa_w1}{}

\item{kappa_eta}{}

\item{kappa_mu}{}

\item{beta}{}

\item{nbd_info}{}

\item{no_of_nbd}{}
}
\description{
Title
}
