% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ICAR_RE.R
\name{W2_sim_Gibbs_componentwise_thr}
\alias{W2_sim_Gibbs_componentwise_thr}
\title{Title}
\usage{
W2_sim_Gibbs_componentwise_thr(
  node_index,
  W2,
  mean_mu_latent,
  kappa_w2,
  kappa_mu,
  nbd_info,
  no_of_nbd
)
}
\arguments{
\item{node_index}{}

\item{W2}{}

\item{kappa_w2}{}

\item{kappa_mu}{}

\item{nbd_info}{}

\item{no_of_nbd}{}

\item{mu_latent}{}
}
\description{
Title
}
